Refactor components for readability.
